content_creation/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
psycopg[binary,pool]>=3.1.0
alembic>=1.13.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
boto3>=1.34.0
langchain-core>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
langdetect>=1.0.9
redis>=5.0.0
sse-starlette>=1.8.0
langchain-mcp-adapters>=0.1.2
aiosqlite>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
testcontainers[postgres]>=4.0.0
locust>=2.20.0
